Factor Nmn explanation into its own file
