+++ demo/src/components/demos/ReadingComprehension.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import { Tooltip, Collapse } from 'antd';
-import styled from 'styled-components';
 import HeatMap from '../HeatMap'
 import { withRouter } from 'react-router-dom';
 import {
@@ -26,9 +24,7 @@ import {
   INPUT_REDUCTION_ATTACKER,
   HOTFLIP_ATTACKER
 } from '../InterpretConstants'
-import NestedHighlight, { withHighlightClickHandling, getHighlightColor } from '../highlight/NestedHighlight';
-import { getHighlightConditionalClasses } from '../highlight/Highlight';
-import FormItem from 'antd/lib/form/FormItem';
+import NmnDropExplanation from './ReadingComprehension/NmnDropExplanation';
 
 const title = "Reading Comprehension"
 
@@ -110,232 +106,6 @@ const NoAnswer = () => {
     </OutputField>
   )
 }
-
-const HighlightTooltipData = props => {
-  const {
-    data,
-    index
-  } = props;
-
-  return (
-    <div>
-      {Object.keys(data)
-        .filter(m => !!data[m].values[index])
-        .map(m => (
-          <div key={m}>{data[m].displayName}: {data[m].values[index]}</div>
-        ))}
-    </div>
-  )
-}
-
-const DEFAULT_FILTER_VALUE = 0.001
-
-class NmnDrop extends React.Component {
-  state = {
-    questionData: {},
-    passageData: {},
-    filterValues: {},
-  }
-
-  componentDidMount() {
-    this.updateData();
-  }
-
-  componentDidUpdate(prevProps) {
-    if (prevProps.programExecution !== this.props.programExecution) {
-      this.updateData();
-    }
-  }
-
-  updateData = () => {
-    const {
-      programExecution,
-      questionTokens,
-      passageTokens,
-      programNestedExpression,
-    } = this.props;
-    const {
-      filterValues
-    } = this.state;
-    const moduleNames = programNestedExpression.flat(Infinity).reverse()
-    const questionData = getHighlightClusters(programExecution, 'question', filterValues, questionTokens, moduleNames)
-    const passageData = getHighlightClusters(programExecution, 'passage', filterValues, passageTokens, moduleNames)
-    this.setState({ questionData, passageData });
-  }
-
-  handleUpdateFilterValue = (name) => (e) => {
-    const filterValue = e.target.value;
-    const filterValues = {
-      ...this.state.filterValues,
-      [name]: filterValue
-    }
-    this.setState({ filterValues }, this.updateData)
-  }
-
-  render () {
-    const { 
-      activeIds,
-      activeDepths,
-      isClicking,
-      selectedId,
-      onMouseDown,
-      onMouseOut,
-      onMouseOver,
-      onMouseUp,
-      programLisp,
-      questionTokens,
-      passageTokens,
-      answer,
-      question,
-    } = this.props
-    const {
-      questionData,
-      passageData,
-    } = this.state;
-    const questionClusters = Object.keys(questionData).reduce((data, d) => {
-      data[d] = questionData[d].clusters;
-      return data;
-    }, {})
-    const passageClusters = Object.keys(passageData).reduce((data, d) => {
-      data[d] = passageData[d].clusters;
-      return data;
-    }, {})
-    const deepestIndex = activeDepths ? activeDepths.depths.indexOf(Math.max(...activeDepths.depths)) : null;
-    return (
-      <section>
-        <OutputField label="Answer">
-          {answer}
-        </OutputField>
-        <OutputField label="Explanation">
-        <QuestionStep>
-            {question}
-          </QuestionStep>
-          <QuestionStep>
-            ↓
-          </QuestionStep>
-          <QuestionStep>
-            Question Parser
-          </QuestionStep>
-          <QuestionStep>
-            ↓
-          </QuestionStep>
-          <CodeQuestionStep>
-            {programLisp}
-          </CodeQuestionStep>
-          <QuestionStep>
-            ↓
-          </QuestionStep>
-          <div>
-            {Object.keys(questionData).map((key, i) => 
-              (
-                <div 
-                  key={i}
-                  style={{ display: 'flex', justifyContent: 'center' }}
-                >
-                  <div 
-                    style={{ 
-                      display: 'unset',
-                      textAlign: 'center',
-                      width: `${(Object.keys(questionData).length - i) / Object.keys(questionData).length * 100}%`
-                    }}
-                    className={getHighlightConditionalClasses({
-                      labelPosition: null,
-                      label: null,
-                      color: getHighlightColor(i),
-                      isClickable: true,
-                      selectedId,
-                      isClicking,
-                      id: key,
-                      activeDepths,
-                      deepestIndex,
-                      activeIds,
-                      children: null,
-                    })}
-                    onMouseDown={onMouseDown ? () => onMouseDown(key, 0) : null}
-                    onMouseOver={onMouseOver ? () => onMouseOver(key) : null}
-                    onMouseOut={onMouseOut ? () => onMouseOut(key) : null}
-                    onMouseUp={onMouseUp ? () => onMouseUp(key) : null}
-                  >
-                    {questionData[key].displayName}
-                  </div>
-                </div>
-              )
-            )}
-          </div>
-          <QuestionStep>
-            ↓
-          </QuestionStep>
-          <QuestionStep>
-            {answer}
-          </QuestionStep>
-        </OutputField>
-        <OutputField label="Question">
-          <NestedHighlight 
-              activeDepths={activeDepths}
-              activeIds={activeIds}
-              clusters={questionClusters}
-              isClickable
-              isClicking={isClicking}
-              labelPosition="bottom"
-              onMouseDown={onMouseDown}
-              onMouseOut={onMouseOut}
-              onMouseOver={onMouseOver}
-              onMouseUp={onMouseUp}
-              selectedId={selectedId}
-              tokens={questionTokens.map((t, i) => <Tooltip title={<HighlightTooltipData index={i} data={questionData} />}>{t}</Tooltip>)}
-            />
-        </OutputField>
-        <OutputField label="Passage">
-          <NestedHighlight
-              activeDepths={activeDepths}
-              activeIds={activeIds}
-              clusters={passageClusters}
-              isClickable
-              isClicking={isClicking}
-              labelPosition="bottom"
-              onMouseDown={onMouseDown}
-              onMouseOut={onMouseOut}
-              onMouseOver={onMouseOver}
-              onMouseUp={onMouseUp}
-              selectedId={selectedId}
-              tokens={passageTokens.map((t, i) => <Tooltip title={<HighlightTooltipData index={i} data={passageData} />}>{t}</Tooltip>)}
-            />
-        </OutputField>
-        <OutputField label="For Model Developers">
-          <Collapse>
-            <Collapse.Panel header="Highlight Sensitivity" key="1">
-              {Object.keys(questionData).map(d => (
-                  <FormItem key={d}>
-                    <label htmlFor="highlight-sensitivity">Adjust Highlight Sensitivity for {d}</label>
-                    <input
-                      id="highlight-sensitivity"
-                      type="range"
-                      min="0.00001"
-                      max="0.01"
-                      step="0.00001"
-                      className="slider"
-                      value={this.state.filterValues[d] || DEFAULT_FILTER_VALUE}
-                      onChange={this.handleUpdateFilterValue(d)}
-                    />
-                  </FormItem>
-              ))}
-            </Collapse.Panel>
-          </Collapse>
-        </OutputField>
-      </section>
-    );
-  }
-}
-
-const QuestionStep = styled.div`
-  text-align: center;
-`;
-
-const CodeQuestionStep = styled(QuestionStep)`
-  font-family: monospace;
-`;
-
-const NmnDropExplanation = withHighlightClickHandling(NmnDrop);
 
 const MultiSpanHighlight = ({original, highlightSpans, highlightStyles}) => {
   if(original && highlightSpans && highlightStyles) {
@@ -443,65 +213,6 @@ const Attacks = ({attackData, attackModel, requestData}) => {
       </Accordion>
     </OutputField>
   )
-}
-
-const setIsEqual = (a, b) => {
-  return Array.from(a).every(value => b.has(value)) && Array.from(b).every(value => a.has(value))
-}
-
-const getHighlightClusters = (programExecution, keyType, filterValues, tokens, moduleNames) => {
-  const initialClusters = moduleNames.reduce((names, name) => {
-    let key = name;
-    // Rename the cluster key if there are duplicates
-    while (names[key]) {
-      key = `${key}_`
-    }
-    names[key] = { 
-      displayName: name,
-      values: [],
-      highlightedIndicies: new Set(),
-      clusters: [],
-    }
-    return names;
-  }, {})
-  const naiveClusters = programExecution.reduce((highlightClusters, e) => {
-    Object.keys(e).forEach(key => {
-      if (e[key][keyType]) {
-        let clusterKey = key
-        while (highlightClusters[clusterKey].values.length !== 0) {
-          clusterKey = `${clusterKey}_`
-        }
-        highlightClusters[clusterKey] = {
-          displayName: key,
-          values: e[key][keyType],
-          highlightedIndicies: new Set(e[key][keyType].map((value, i) => value > (filterValues[clusterKey] || DEFAULT_FILTER_VALUE) ? i : undefined).filter(i => i)),
-          clusters: [],
-        }
-      }
-    })
-    return highlightClusters;
-  }, initialClusters);
-  const modules = Object.keys(naiveClusters)
-  let prevModulesWithHighlight = new Set()
-  for (let i = 0; i < tokens.length; i++) {
-    const modulesWithHighlight = new Set()
-    for (let n = 0; n < modules.length; n++) {
-      if (naiveClusters[modules[n]].highlightedIndicies.has(i)) {
-        modulesWithHighlight.add(modules[n])
-      }
-    }
-    if (setIsEqual(prevModulesWithHighlight, modulesWithHighlight)) {
-      modulesWithHighlight.forEach(m => {
-        naiveClusters[m].clusters[naiveClusters[m].clusters.length - 1][1] = i;
-      })
-    } else {
-      modulesWithHighlight.forEach(m => {
-        naiveClusters[m].clusters.push([i, i])
-      })
-    }
-    prevModulesWithHighlight = modulesWithHighlight;
-  }
-  return naiveClusters;
 }
 
 const AnswerByType = ({ responseData, requestData, interpretData, interpretModel, attackData, attackModel}) => {
